More friendly option
Use a service, couldn't found a solution that I liked without using serviceprovider out of startup but this one is fine
